+++ src/sc_types.h
@@ -43,7 +43,7 @@ enum insn_type {
 
 #define INSN_DISASM_MAX_LEN 64
 typedef struct insn_info {
-	unsigned int len;
+	unsigned long len;
 	unsigned long insn;
 	char disasm[INSN_DISASM_MAX_LEN];
 } insn_info_t;
